I think? it works can't test rn

+++ src/main/java/com/gregtechceu/gtceu/api/machine/multiblock/MultiblockDisplayText.java
@@ -205,7 +205,7 @@ public class MultiblockDisplayText {
          * <br>
          * Added if the structure is formed and if the max CWU/t is greater than zero.
          */
-        public Builder addComputationUsageLine(int maxCWUt) {
+        public Builder addComputationUsageLine(long maxCWUt) {
             if (!isStructureFormed)
                 return this;
             if (maxCWUt > 0) {
